+++ Nbm/src/test/resources/projForTest/src/diagnostics/checkJdkTypesVisible.kt
@@ -1,0 +1,3 @@
+package diagnostics
+
+class MyException(msg: String) : Exception(msg)